Pull client and server into own files
