+++ packages/iov-transport/package.json
@@ -25,7 +25,7 @@
     "test": "yarn build-or-skip && yarn test-node",
     "prebuild": "yarn format && yarn lint",
     "preprocess-types": "rm -r ./types/*",
-    "postprocess-types": "echo hi",
+    "postprocess-types": "rm -r ./types/*.spec.d.ts",
     "build": "yarn preprocess-types && tsc --declarationDir ./types && yarn postprocess-types",
     "build-or-skip": "[ -n \"$SKIP_BUILD\" ] || yarn build",
     "pack-web": "yarn build-or-skip && webpack --mode development --config webpack.web.config.js"
