+++ packages/iov-transport/src/connection.ts
@@ -10,17 +10,7 @@ This can run over many transport layers.
 
 import { Listener, Producer, Stream } from "xstream";
 
-import {
-  envelope,
-  Event,
-  format,
-  Message,
-  MessageKind,
-  randomId,
-  RequestMessage,
-  ResponseMessage,
-  SubscribeMessage,
-} from "./messages";
+import { Event, Message } from "./messages";
 
 // A connector is anything that can generate a connection (client)
 export type Connector = () => Connection;
@@ -34,7 +24,7 @@ export interface Connection {
 }
 
 // SendProducer<T> allows us to send events to a listener
-class SendProducer<T> implements Producer<T> {
+export class SendProducer<T> implements Producer<T> {
   // tslint:disable-next-line:readonly-keyword
   private listener: Listener<T> | undefined;
 
@@ -58,8 +48,8 @@ class SendProducer<T> implements Producer<T> {
   }
 }
 
-type EventProducer = SendProducer<Event>;
-type MessageProducer = SendProducer<Message>;
+export type EventProducer = SendProducer<Event>;
+export type MessageProducer = SendProducer<Message>;
 
 // localConnectionPair returns a pair of connections that can be
 // pass to Client and Server for them to communicate.
@@ -77,155 +67,3 @@ const localConnection = (client: MessageProducer, server: MessageProducer): Conn
   receive: Stream.create(server),
   disconnect: () => client.stop(),
 });
-
-// Promiser holds the callbacks to later resolve Promise from different
-// control flow
-interface Promiser {
-  readonly resolve: (value: any) => void;
-  readonly reject: (err: any) => void;
-}
-
-// Client sends requests to a connection and gets responses,
-// possibly out of order
-export class Client {
-  private readonly connection: Connection;
-  private readonly resolvers: Map<string, Promiser>;
-  private readonly streams: Map<string, EventProducer>;
-
-  constructor(connection: Connection) {
-    this.connection = connection;
-    this.resolvers = new Map<string, Promiser>();
-    this.streams = new Map<string, EventProducer>();
-    this.listen();
-  }
-
-  public request(method: string, params: any): Promise<any> {
-    const req: RequestMessage = {
-      ...envelope(),
-      kind: MessageKind.REQUEST,
-      method,
-      params,
-    };
-    this.connection.send(req);
-    return new Promise<any>((resolve, reject) => {
-      // make sure nothing already registered
-      if (this.resolvers.get(req.id)) {
-        reject(`Cannot reuse ${req.id}, request currently pending`);
-      } else {
-        this.resolvers.set(req.id, { resolve, reject });
-      }
-    });
-  }
-
-  public subscribe(query: string): Stream<Event> {
-    const subscriptionId = randomId();
-
-    const sub: SubscribeMessage = {
-      ...envelope(),
-      kind: MessageKind.REQUEST,
-      method: "subscribe",
-      params: {
-        query,
-        subscriptionId,
-      },
-    };
-    this.connection.send(sub);
-    const producer = new SendProducer<Event>();
-    // TODO: somehow store this to unsubscribe if subscription request
-    // above returns an error
-    // this.resolvers.set(sub.id, ...)
-    this.streams.set(subscriptionId, producer);
-    return Stream.create(producer);
-  }
-
-  private listen(): void {
-    const subscription = this.connection.receive.subscribe({
-      next: (msg: Message) => this.handleMessage(msg),
-      error: x => {
-        throw x;
-      },
-      complete: () => {
-        subscription.unsubscribe();
-        this.connection.disconnect();
-      },
-    });
-  }
-
-  private handleMessage(msg: Message): void {
-    if (msg.format !== format) {
-      const text = JSON.stringify(msg);
-      throw new Error(`Invalid message: ${text}`);
-    }
-
-    switch (msg.kind) {
-      case MessageKind.RESPONSE:
-      case MessageKind.ERROR:
-        const res = this.resolvers.get(msg.id);
-        if (!res) {
-          throw new Error(`Unknown response ${msg.id}`);
-        } else {
-          msg.kind === MessageKind.RESPONSE ? res.resolve(msg.result) : res.reject(msg.error);
-          this.resolvers.delete(msg.id);
-        }
-        break;
-      case MessageKind.EVENT:
-        const prod = this.streams.get(msg.id);
-        if (!prod) {
-          throw new Error(`Unknown event ${msg.id}`);
-        } else {
-          prod.send(msg.event);
-        }
-    }
-  }
-}
-
-// Handler is the mirror of client, but different names,
-// signaling it is doing the work, not requesting it
-export interface Handler {
-  readonly handleRequest: (method: string, params: any) => Promise<any>;
-  readonly handleSubscribe: (query: string) => Stream<Event>;
-}
-
-export class Server {
-  private readonly connection: Connection;
-  private readonly handler: Handler;
-
-  constructor(connection: Connection, handler: Handler) {
-    this.connection = connection;
-    this.handler = handler;
-    this.listen();
-  }
-
-  private listen(): void {
-    const subscription = this.connection.receive.subscribe({
-      next: (msg: Message) => this.handleMessage(msg),
-      error: x => {
-        throw x;
-      },
-      complete: () => {
-        subscription.unsubscribe();
-        this.connection.disconnect();
-      },
-    });
-  }
-
-  private async handleMessage(msg: Message): Promise<void> {
-    if (msg.format !== format) {
-      const text = JSON.stringify(msg);
-      throw new Error(`Invalid message: ${text}`);
-    }
-
-    // only handle requests, ignore others
-    if (msg.kind === MessageKind.REQUEST) {
-      // TODO: handle subscriptions specially
-      const result = await this.handler.handleRequest(msg.method, msg.params);
-      const response: ResponseMessage = {
-        format,
-        id: msg.id,
-        kind: MessageKind.RESPONSE,
-        result,
-      };
-      this.connection.send(response);
-    }
-  }
-}
