+++ packages/iov-transport/types/connection.d.ts
@@ -1,4 +1,4 @@
-import { Stream } from "xstream";
+import { Listener, Producer, Stream } from "xstream";
 import { Event, Message } from "./messages";
 export declare type Connector = () => Connection;
 export interface Connection {
@@ -6,25 +6,12 @@ export interface Connection {
     readonly receive: Stream<Message>;
     readonly disconnect: () => void;
 }
+export declare class SendProducer<T> implements Producer<T> {
+    private listener;
+    start(listener: Listener<T>): void;
+    stop(): void;
+    send(msg: T): void;
+}
+export declare type EventProducer = SendProducer<Event>;
+export declare type MessageProducer = SendProducer<Message>;
 export declare const localConnectionPair: () => [Connection, Connection];
-export declare class Client {
-    private readonly connection;
-    private readonly resolvers;
-    private readonly streams;
-    constructor(connection: Connection);
-    request(method: string, params: any): Promise<any>;
-    subscribe(query: string): Stream<Event>;
-    private listen;
-    private handleMessage;
-}
-export interface Handler {
-    readonly handleRequest: (method: string, params: any) => Promise<any>;
-    readonly handleSubscribe: (query: string) => Stream<Event>;
-}
-export declare class Server {
-    private readonly connection;
-    private readonly handler;
-    constructor(connection: Connection, handler: Handler);
-    private listen;
-    private handleMessage;
-}
