+++ packages/iov-transport/types/index.d.ts
@@ -1,2 +1,4 @@
+export * from "./client";
 export * from "./connection";
 export * from "./messages";
+export * from "./server";
